Update README.md
This folder contains a number of algorithms that I created as a student. The False Position root-finding method uses numerical techniques to estimate the roots of a function. The LU Decomposition decomposes matrices to solve systems of linear equations. Linear Regression is a statistical technique used to model the relationship between two variables. Finally, Simpsons Integration is a numerical method used to approximate the area under a curve.

This folder provides researchers with a variety of tools unique to their respective fields, each of which has distinct features that make them useful in their respective fields.