Update False Position
Example:
Let f(x) = x^2 - 5 and we want to find its root with an error of 0.01. We can start with xl = 2 and xu = 3, where xl and xu must have different signs if we evaluate them.
Once the code is executed, we will be able to determine the root of the function with a precision of 0.01. The result we obtain will be around 2.236 where the function equals zero.