Update Linear Regression
To use this code, you must enter two sets of data: x and y values. The algorithm will then produce three results: the line's slope, intercept, and coefficient of determination.

The slope of the line, which specifies the direction and steepness of the line, will be the code's initial output. The intercept will be the second output, indicating where the line crosses the y-axis. Finally, the coefficient of determination will be output, which represents how well the line fits the data.