Update Simpsons Integration
This code aims to compute the numerical value of an integral using Simpson's 1/3 Rule. To run this code, you must supply two input vectors: a vector of equally spaced independent variables (x) and a vector of function values (y). The numerical integral will then be calculated and output by the code and its output will be the estimated value of the numerical integral (I).
Example:
Let x = [0 1 2] and y = [3 4 5]. By running the code, you will obtain the numerical integration of the function, which will be approximately equal to 7.67.